+++ sdk/identity/Azure.Identity/tests/SharedTokenCacheCredentialTests.cs
@@ -401,5 +401,58 @@ namespace Azure.Identity.Tests
 
             await Task.CompletedTask;
         }
+
+        [Test]
+        public async Task MatchAnySingleTenantIdWithEnableGuestTenantAuthentication()
+        {
+            string expToken = Guid.NewGuid().ToString();
+            DateTimeOffset expExpiresOn = DateTimeOffset.UtcNow.AddMinutes(5);
+            string tenantId = Guid.NewGuid().ToString();
+            var mockMsalClient = new MockMsalPublicClient
+            {
+                Accounts = new List<IAccount> { new MockAccount("mockuser@mockdomain.com", Guid.NewGuid().ToString()) },
+                SilentAuthFactory = (_) => { return AuthenticationResultFactory.Create(accessToken: expToken, expiresOn: expExpiresOn); }
+            };
+
+            var credential = InstrumentClient(new SharedTokenCacheCredential(tenantId, null, new SharedTokenCacheCredentialOptions { EnableGuestTenantAuthentication = true }, null, mockMsalClient));
+
+            AccessToken token = await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Default));
+
+            Assert.AreEqual(expToken, token.Token);
+
+            Assert.AreEqual(expExpiresOn, token.ExpiresOn);
+        }
+
+        [Test]
+        public async Task MatchAnyTenantIdWithEnableGuestTenantAuthenticationAndUsername()
+        {
+            string expToken = Guid.NewGuid().ToString();
+            DateTimeOffset expExpiresOn = DateTimeOffset.UtcNow.AddMinutes(5);
+            string tenantId = Guid.NewGuid().ToString();
+            var mockMsalClient = new MockMsalPublicClient
+            {
+                Accounts = new List<IAccount> { new MockAccount("mockuser@mockdomain.com", Guid.NewGuid().ToString()), new MockAccount("fakeuser@fakedomain.com", Guid.NewGuid().ToString()) },
+                SilentAuthFactory = (_) => { return AuthenticationResultFactory.Create(accessToken: expToken, expiresOn: expExpiresOn); }
+            };
+
+
+            var credential = InstrumentClient(new SharedTokenCacheCredential(tenantId, "mockuser@mockdomain.com", new SharedTokenCacheCredentialOptions { EnableGuestTenantAuthentication = true }, null, mockMsalClient));
+
+            AccessToken token = await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Default));
+
+            Assert.AreEqual(expToken, token.Token);
+
+            Assert.AreEqual(expExpiresOn, token.ExpiresOn);
+        }
+
+        [Test]
+        public void ValidateClientIdSetOnMsalClient()
+        {
+            var clientId = Guid.NewGuid().ToString();
+
+            var credential = new SharedTokenCacheCredential(new SharedTokenCacheCredentialOptions { ClientId = clientId });
+
+            Assert.AreEqual(clientId, credential.Client.ClientId);
+        }
     }
 }